decouple TrainingController
