+++ src/Controller/TrainingController.php
@@ -4,10 +4,8 @@ declare(strict_types=1);
 
 namespace Pehapkari\Controller;
 
-use Nette\Utils\DateTime;
 use Pehapkari\Registration\Repository\TrainingRegistrationRepository;
 use Pehapkari\Repository\TrainerRepository;
-use Pehapkari\Training\Entity\Training;
 use Pehapkari\Training\Repository\TrainingFeedbackRepository;
 use Pehapkari\Training\Repository\TrainingRepository;
 use Pehapkari\Training\Repository\TrainingTermRepository;
@@ -80,38 +78,5 @@ final class TrainingController extends AbstractController
 
             'trainer_count' => count($this->trainerRepository->fetchAll()),
         ]);
-    }
-
-    /**
-     * @Route(path="/kurz/{slug}", name="training_detail")
-     */
-    public function detail(Training $training): Response
-    {
-        $averageRating = $this->trainingFeedbackRepository->getAverageTrainingRating($training);
-
-        return $this->render('training/training_detail.twig', [
-            'training' => $training,
-            'training_term' => $training->getNearestTerm(),
-            'trainer' => $training->getTrainer(),
-            'should_display_deadline' => $this->shouldDisplayDeadline($training),
-
-            // rating
-            'feedbacks' => $training->getPublicFeedbacks(),
-            'average_training_rating' => $averageRating,
-            'average_training_rating_stars' => round($averageRating, 0),
-        ]);
-    }
-
-    private function shouldDisplayDeadline(Training $training): bool
-    {
-        $nearestTerm = $training->getNearestTerm();
-        if ($nearestTerm === null) {
-            return false;
-        }
-
-        // show only on nearest X days
-        $weekBackDateTime = (new DateTime())->modify(' - 14 days');
-
-        return $nearestTerm->getDeadlineDateTime() < $weekBackDateTime;
     }
 }
