+++ src/Controller/TrainingDetailController.php
@@ -1,0 +1,51 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Pehapkari\Controller;
+
+use Nette\Utils\DateTime;
+use Pehapkari\Training\Entity\Training;
+use Pehapkari\Training\Repository\TrainingFeedbackRepository;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Routing\Annotation\Route;
+
+final class TrainingDetailController extends AbstractController
+{
+    /**
+     * @Route(path="/kurz/{slug}", name="training_detail")
+     */
+    public function detail(Training $training, TrainingFeedbackRepository $trainingFeedbackRepository): Response
+    {
+        $averageRating = $trainingFeedbackRepository->getAverageTrainingRating($training);
+
+        dump($training);
+        die;
+
+        return $this->render('training/training_detail.twig', [
+            'training' => $training,
+            'training_term' => $training->getNearestTerm(),
+            'trainer' => $training->getTrainer(),
+            'should_display_deadline' => $this->shouldDisplayDeadline($training),
+
+            // rating
+            'feedbacks' => $training->getPublicFeedbacks(),
+            'average_training_rating' => $averageRating,
+            'average_training_rating_stars' => round($averageRating, 0),
+        ]);
+    }
+
+    private function shouldDisplayDeadline(Training $training): bool
+    {
+        $nearestTerm = $training->getNearestTerm();
+        if ($nearestTerm === null) {
+            return false;
+        }
+
+        // show only on nearest X days
+        $weekBackDateTime = (new DateTime())->modify(' - 14 days');
+
+        return $nearestTerm->getDeadlineDateTime() < $weekBackDateTime;
+    }
+}
